feat: add map to progress and refine filtering behavior
