+++ apps/web/src/components/popover.tsx
@@ -11,12 +11,22 @@ function PopoverTrigger({ ...props }: PopoverPrimitive.Trigger.Props) {
   return <PopoverPrimitive.Trigger data-slot="popover-trigger" {...props} />;
 }
 
+// Move focus into the popup only on keyboard open — mouse opens leave
+// focus alone, so the first tabbable control (e.g. a Toggle) never gets
+// auto-focused where a following Escape/Space would visibly act on it.
+// Same convention as DropdownMenuContent's finalFocus; callsites can
+// override either prop.
+const INITIAL_FOCUS_KEYBOARD_ONLY = (openType: string): boolean => openType === "keyboard";
+const FINAL_FOCUS_KEYBOARD_ONLY = (closeType: string): boolean => closeType === "keyboard";
+
 function PopoverContent({
   className,
   align = "center",
   alignOffset = 0,
   side = "bottom",
   sideOffset = 4,
+  initialFocus = INITIAL_FOCUS_KEYBOARD_ONLY,
+  finalFocus = FINAL_FOCUS_KEYBOARD_ONLY,
   ...props
 }: PopoverPrimitive.Popup.Props &
   Pick<PopoverPrimitive.Positioner.Props, "align" | "alignOffset" | "side" | "sideOffset">) {
@@ -31,6 +41,8 @@ function PopoverContent({
       >
         <PopoverPrimitive.Popup
           data-slot="popover-content"
+          initialFocus={initialFocus}
+          finalFocus={finalFocus}
           className={cn(
             "z-50 flex w-72 origin-(--transform-origin) flex-col gap-2.5 rounded-lg bg-popover p-2.5 text-sm text-popover-foreground shadow-md ring-1 ring-foreground/10 outline-hidden duration-100 data-[side=bottom]:slide-in-from-top-2 data-[side=inline-end]:slide-in-from-left-2 data-[side=inline-start]:slide-in-from-right-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 data-open:animate-in data-open:fade-in-0 data-open:zoom-in-95 data-closed:animate-out data-closed:fade-out-0 data-closed:zoom-out-95",
             className,
