+++ apps/web/src/lib/geometry.ts
@@ -1,3 +1,30 @@
+// Bounding box over features' coordinates, as [minLng, minLat, maxLng,
+// maxLat] for the map's fitBounds; null when there are no coordinates.
+export function bboxOfFeatures(
+  features: GeoJSON.Feature[],
+): [number, number, number, number] | null {
+  let minLng = Infinity;
+  let minLat = Infinity;
+  let maxLng = -Infinity;
+  let maxLat = -Infinity;
+  const visit = (coords: unknown) => {
+    if (!Array.isArray(coords)) return;
+    if (typeof coords[0] === "number" && typeof coords[1] === "number") {
+      minLng = Math.min(minLng, coords[0] as number);
+      maxLng = Math.max(maxLng, coords[0] as number);
+      minLat = Math.min(minLat, coords[1] as number);
+      maxLat = Math.max(maxLat, coords[1] as number);
+      return;
+    }
+    for (const c of coords) visit(c);
+  };
+  for (const f of features) {
+    if ("coordinates" in f.geometry) visit(f.geometry.coordinates);
+  }
+  if (!Number.isFinite(minLng)) return null;
+  return [minLng, minLat, maxLng, maxLat];
+}
+
 // Ray-casting point-in-polygon. Works in any coordinate system —
 // callers use either lng/lat or screen coords.
 export function pointInPolygon(point: [number, number], polygon: Array<[number, number]>): boolean {
